refactor: eventScheduler all in utils
